TS: extracting SequenceEditor from IPScan/Search
