+++ package.json
@@ -153,6 +153,7 @@
     "@types/color-hash": "^1.0.5",
     "@types/d3-scale": "4.0.8",
     "@types/dompurify": "3.0.5",
+    "@types/draft-js": "0.11.18",
     "@types/gtag.js": "0.0.18",
     "@types/lodash-es": "4.17.12",
     "@types/react-dom": "18.2.18",
