+++ src/components/IPScan/Search/index.js
@@ -4,15 +4,6 @@ import { connect } from 'react-redux';
 import { createSelector } from 'reselect';
 import { Helmet } from 'react-helmet-async';
 
-import {
-  Editor,
-  EditorState,
-  ContentState,
-  CompositeDecorator,
-  Modifier,
-  convertToRaw,
-} from 'draft-js';
-
 import AdvancedOptions from 'components/IPScan/AdvancedOptions';
 import Redirect from 'components/generic/Redirect';
 import Link from 'components/generic/Link';
@@ -20,7 +11,6 @@ import Link from 'components/generic/Link';
 import Callout from 'components/SimpleCommonComponents/Callout';
 // $FlowFixMe
 import Button from 'components/SimpleCommonComponents/Button';
-import getId from 'utils/cheap-unique-id';
 
 import { createJob, goToCustomLocation } from 'actions/creators';
 import loadable from 'higherOrder/loadable';
@@ -40,6 +30,7 @@ import buttonCSS from 'components/SimpleCommonComponents/Button/style.css';
 import fonts from 'EBI-Icon-fonts/fonts.css';
 
 import example from './example.fasta';
+import SequenceInput from './SequenceInput';
 
 const f = foundationPartial(
   interproTheme,
@@ -57,58 +48,6 @@ const SchemaOrgData = loadable({
   loading: () => null,
 });
 
-const strategy = (re) => (block, cb) => {
-  const text = block.getText();
-  let match;
-  while ((match = re.exec(text))) {
-    cb(match.index, match.index + match[0].length);
-  }
-};
-const doesPreviousBlockHasContent = (block, contentState) => {
-  const before = contentState.getBlockBefore(block.key);
-  if (!before) return true;
-  const trimmed = before.getText().trim();
-  if (trimmed === '' || trimmed.startsWith(';'))
-    return doesPreviousBlockHasContent(before, contentState);
-  return !trimmed.startsWith('>');
-};
-const commentsStrategy =
-  (re, forErrors = false) =>
-  (block, cb, contentState) => {
-    const text = block.getText();
-    let match;
-    while ((match = re.exec(text))) {
-      const prevSequenceOK = doesPreviousBlockHasContent(block, contentState);
-      if (!forErrors && (prevSequenceOK || text.trim().startsWith(';')))
-        cb(match.index, match.index + match[0].length);
-      if (forErrors && !prevSequenceOK)
-        cb(match.index, match.index + match[0].length);
-    }
-  };
-/*:: type ScanProps = {
-      offsetKey: string,
-      children: any
-  }*/
-const classedSpan = (className) => {
-  class Span extends PureComponent /*:: <ScanProps> */ {
-    static propTypes = {
-      offsetKey: T.string.isRequired,
-      children: T.any,
-    };
-
-    render() {
-      const { offsetKey, children } = this.props;
-      return (
-        <span className={className} data-offset-key={offsetKey}>
-          {children}
-        </span>
-      );
-    }
-  }
-
-  return Span;
-};
-
 const checkedSelectorFor = (x) => `input[name="${x}"]:checked`;
 
 const getCheckedApplications = (form) =>
@@ -121,164 +60,6 @@ const isXChecked = (x) => (form) => !!form.querySelector(checkedSelectorFor(x));
 
 const isStayChecked = isXChecked('stay');
 
-const headerRE = /^\s*>.*$/m;
-const IUPACProtRE = /^[a-z\s]*$/im;
-
-const trimSequenceLines = (lines) =>
-  lines
-    .map((l) => l.trim()) // trimming
-    .map((l) => (l.startsWith(';') ? false : l)) // removing comments
-    .filter(Boolean);
-
-export const checkValidity = (lines) => {
-  let numberOfSequences = 0;
-  let gotContentYet = false;
-  const trimmedLines = trimSequenceLines(lines);
-  if (!headerRE.test(trimmedLines[0]) && !IUPACProtRE.test(trimmedLines[0]))
-    return false;
-  for (const line of trimmedLines) {
-    if (headerRE.test(line)) {
-      if (numberOfSequences > 0 && !gotContentYet) return false;
-      gotContentYet = false;
-      numberOfSequences++;
-    } else if (IUPACProtRE.test(line)) {
-      gotContentYet = true;
-    } else {
-      return false;
-    }
-  }
-  return numberOfSequences <= MAX_NUMBER_OF_SEQUENCES && gotContentYet;
-};
-
-const MIN_LENGTH = 3;
-
-export const isTooShort = (lines) => {
-  let count = 0;
-  let firstLine = true;
-  const trimmedLines = trimSequenceLines(lines);
-  for (const line of trimmedLines) {
-    if (headerRE.test(line)) {
-      if (!firstLine && count < MIN_LENGTH) return true;
-      count = 0;
-    } else {
-      count += line.trim().length;
-    }
-    firstLine = false;
-  }
-  return count < MIN_LENGTH;
-};
-
-const hasHeaderIssues = (lines) => {
-  const trimmedLines = trimSequenceLines(lines);
-  const isFirstLineAComment =
-    trimmedLines.length && trimmedLines[0].startsWith('>');
-  const numberOfHeaders = trimmedLines.filter((l) => l.startsWith('>')).length;
-  if (numberOfHeaders === 0) return false;
-  if (numberOfHeaders === 1 && isFirstLineAComment) return false;
-  for (let x = 1; x < lines.length; x++) {
-    if (lines[x - 1].startsWith('>') && lines[x].startsWith('>')) return true; // it has 2 consecutive headers
-  }
-  return false;
-};
-const hasTooManySequences = (lines) => {
-  const trimmedLines = trimSequenceLines(lines);
-  return (
-    trimmedLines.filter((line) => line.startsWith('>')).length >
-    MAX_NUMBER_OF_SEQUENCES
-  );
-};
-const compositeDecorator = new CompositeDecorator([
-  {
-    strategy: commentsStrategy(/^\s*[>;].*$/gm),
-    component: classedSpan(f('comment')),
-  },
-  {
-    strategy: commentsStrategy(/^\s*[>].*$/gm, true),
-    component: classedSpan(f('invalid-comment')),
-  },
-  {
-    strategy: strategy(/[^a-z]+/gi),
-    component: classedSpan(f('invalid-letter')),
-  },
-]);
-const hasContentInNextLine = (lines, i) => {
-  if (i + 1 >= lines.length) return false;
-  const line = lines[i + 1].text.trim();
-  if (line.startsWith(';') || line === '')
-    return hasContentInNextLine(lines, i + 1);
-  if (line.startsWith('>')) return false;
-  return true;
-};
-export const cleanUp = (blocks) => {
-  return blocks
-    .map(({ text }, i, lines) => {
-      const line = text.trim();
-      if (!line) return null;
-      if (line.startsWith(';')) return line;
-      if (line.startsWith('>'))
-        return hasContentInNextLine(lines, i) ? line : null;
-
-      return line.replace(/[^a-z\s]/gi, '').trim();
-    })
-    .filter(Boolean)
-    .join('\n');
-};
-
-const InfoMessages = ({ valid, tooShort, tooMany, headerIssues }) => {
-  return (
-    <div className={f('text-right')}>
-      {!valid && (
-        <div>
-          {tooShort
-            ? 'There is a header without content. '
-            : 'The sequence has invalid characters. '}
-          <span role="img" aria-label="warning">
-            ⚠️
-          </span>
-        </div>
-      )}
-      {valid && tooShort && (
-        <div>
-          The sequence is too short.{' '}
-          <span role="img" aria-label="warning">
-            ⚠️
-          </span>
-        </div>
-      )}
-      {tooMany && (
-        <div>
-          There are too many sequences. The maximum allowed is{' '}
-          {MAX_NUMBER_OF_SEQUENCES}.{' '}
-          <span role="img" aria-label="warning">
-            ⚠️
-          </span>
-        </div>
-      )}
-      {headerIssues && (
-        <div>
-          There are issues with the headers.{' '}
-          <span role="img" aria-label="warning">
-            ⚠️
-          </span>
-        </div>
-      )}
-      {valid && !tooShort && !headerIssues && (
-        <div>
-          Valid Sequence.{' '}
-          <span role="img" aria-label="warning">
-            ✅
-          </span>
-        </div>
-      )}
-    </div>
-  );
-};
-InfoMessages.propTypes = {
-  valid: T.bool.isRequired,
-  tooShort: T.bool.isRequired,
-  tooMany: T.bool.isRequired,
-  headerIssues: T.bool.isRequired,
-};
 /*:: type Props = {
   createJob: function,
   goToCustomLocation: function,
@@ -313,53 +94,28 @@ export class IPScanSearch extends PureComponent /*:: <Props, State> */ {
   constructor(props /*: Props */) {
     super(props);
 
-    let editorState;
-    let valid = true;
-    let tooShort = true;
-    let tooMany = false;
-    let headerIssues = false;
     let initialAdvancedOptions = null;
-    if (props.value) {
-      editorState = EditorState.createWithContent(
-        ContentState.createFromText(decodeURIComponent(props.value)),
-        compositeDecorator,
-      );
-      const lines = convertToRaw(editorState.getCurrentContent()).blocks.map(
-        (block) => block.text,
-      );
-      valid = checkValidity(lines);
-      tooShort = isTooShort(lines);
-      tooMany = hasTooManySequences(lines);
-      headerIssues = hasHeaderIssues(lines);
-    } else {
-      editorState = EditorState.createEmpty(compositeDecorator);
-    }
+
     if (props.search) {
       initialAdvancedOptions = props.search;
     }
     this.state = {
-      editorState,
-      valid,
-      tooShort,
-      tooMany,
-      headerIssues,
       title: null,
       initialAdvancedOptions,
       submittedJob: null,
+      sequenceChecks: {},
     };
 
     this._formRef = React.createRef();
     this._editorRef = React.createRef();
+    if (this.props.value) this.sequenceToSet = this.props.value;
   }
-
-  _getTitle = (lines) => {
-    if (lines.length) {
-      if (lines[0].startsWith('>')) {
-        return lines[0].split('>')[1];
-      }
+  componentDidUpdate() {
+    if (this.sequenceToSet && this._editorRef.current) {
+      this._handleReset(this.sequenceToSet);
+      this.sequenceToSet = null;
     }
-    return null;
-  };
+  }
 
   _changeTitle = () => {
     const newTitle = this._formRef.current
@@ -382,41 +138,23 @@ export class IPScanSearch extends PureComponent /*:: <Props, State> */ {
       }
     }
 
-    // let fileTitle = '';
-    // if (typeof text === 'string') {
-    // const lines = text.split(/\n/);
-    // fileTitle = this._getTitle(lines);
-    // }
-
     this.setState(
       {
-        editorState:
-          text && typeof text === 'string'
-            ? EditorState.createWithContent(
-                ContentState.createFromText(text),
-                compositeDecorator,
-              )
-            : EditorState.createEmpty(compositeDecorator),
-        valid: true,
-        tooShort: true,
-        tooMany: false,
-        headerIssues: false,
         dragging: false,
         uploading: false,
         title: null,
         submittedJob: null,
       },
-      this._focusEditor,
+      this._editorRef.current.focusEditor(),
     );
+    this._editorRef.current.reset(typeof text === 'string' ? text : undefined);
   };
 
   _handleSubmit = (event) => {
     event.preventDefault();
     if (!this._formRef.current) return;
-    const lines = convertToRaw(
-      this.state.editorState.getCurrentContent(),
-    ).blocks.map((block) => block.text);
-    if (!lines.length) return;
+    const editorContent = this._editorRef.current.getContent();
+    if (!editorContent) return;
     const localID = id(`internal-${Date.now()}`);
     this.props.createJob({
       metadata: {
@@ -425,7 +163,7 @@ export class IPScanSearch extends PureComponent /*:: <Props, State> */ {
         type: 'InterProScan',
       },
       data: {
-        input: lines.join('\n'),
+        input: editorContent,
         applications: getCheckedApplications(this._formRef.current),
       },
     });
@@ -458,10 +196,7 @@ export class IPScanSearch extends PureComponent /*:: <Props, State> */ {
 
   _loadExample = () => this._handleReset(example);
 
-  _cleanUp = () =>
-    this._handleReset(
-      cleanUp(convertToRaw(this.state.editorState.getCurrentContent()).blocks),
-    );
+  _cleanUp = () => this._editorRef.current.cleanUp();
 
   _handleDroppedFiles = blockEvent(
     ({
@@ -480,65 +215,6 @@ export class IPScanSearch extends PureComponent /*:: <Props, State> */ {
     target.value = null;
   };
 
-  _handlePastedText = (pasted) => {
-    const blockMap = ContentState.createFromText(pasted).getBlockMap();
-    const editorState = EditorState.push(
-      this.state.editorState,
-      Modifier.replaceWithFragment(
-        this.state.editorState.getCurrentContent(),
-        this.state.editorState.getSelection(),
-        blockMap,
-      ),
-      'insert-fragment',
-    );
-    this._handleChange(editorState);
-    return true;
-  };
-
-  _focusEditor = () => {
-    if (this._editorRef.current) this._editorRef.current.focus();
-  };
-
-  _addFastAHeaderIfNeeded = (editorState, lines) => {
-    const minLengthForHeader = 3;
-    const currentContent = editorState.getCurrentContent();
-    if (currentContent.hasText()) {
-      const firstLine = (lines?.[0] || '').trim();
-      const hasHeader = firstLine.startsWith('>');
-      if (!hasHeader && firstLine.length > minLengthForHeader) {
-        const newHeader = `Sequence title ${getId()}`;
-        const header = `> ${newHeader}`;
-        // this.setState({ title: newHeader });
-        lines.splice(0, 0, header);
-        const newState = EditorState.createWithContent(
-          ContentState.createFromText(lines.join('\n')),
-          compositeDecorator,
-        );
-        return EditorState.moveFocusToEnd(newState);
-      }
-    }
-    return null;
-  };
-
-  _handleChange = (editorState) => {
-    const lines = convertToRaw(editorState.getCurrentContent()).blocks.map(
-      (block) => block.text,
-    );
-    const stateWithHeader = this._addFastAHeaderIfNeeded(editorState, lines);
-    if (stateWithHeader) {
-      this._handleChange(stateWithHeader);
-      return;
-    }
-    this.setState({
-      editorState,
-      valid: checkValidity(lines),
-      tooShort: isTooShort(lines),
-      tooMany: hasTooManySequences(lines),
-      headerIssues: hasHeaderIssues(lines),
-      // title: this._getTitle(lines),
-    });
-  };
-
   render() {
     // If we had a value, we used it in the constructor
     // But we don't want to have it in the url, so remove the value from it.
@@ -553,8 +229,7 @@ export class IPScanSearch extends PureComponent /*:: <Props, State> */ {
           }}
         />
       );
-    const { editorState, valid, tooShort, tooMany, headerIssues, dragging } =
-      this.state;
+    const { dragging } = this.state;
     return (
       <div className={f('row', 'margin-bottom-medium')}>
         <div className={f('large-12', 'columns')}>
@@ -635,41 +310,13 @@ export class IPScanSearch extends PureComponent /*:: <Props, State> */ {
                         </p>
                       </div>
                     )}
-                    <div
-                      onClick={this._focusEditor}
-                      onKeyPress={this._focusEditor}
-                      role="presentation"
-                    >
-                      <div
-                        type="text"
-                        className={f('editor', {
-                          'invalid-block':
-                            !valid && editorState.getCurrentContent().hasText(),
-                          'valid-block':
-                            valid &&
-                            editorState.getCurrentContent().hasText() &&
-                            !tooShort &&
-                            !headerIssues,
-                        })}
-                      >
-                        <Editor
-                          placeholder="Enter your sequence"
-                          editorState={editorState}
-                          handleDroppedFiles={this._handleDroppedFiles}
-                          onChange={this._handleChange}
-                          handlePastedText={this._handlePastedText}
-                          ref={this._editorRef}
-                        />
-                      </div>
-                    </div>
-                    {editorState.getCurrentContent().hasText() && (
-                      <InfoMessages
-                        valid={valid}
-                        tooShort={tooShort}
-                        headerIssues={headerIssues}
-                        tooMany={tooMany}
-                      />
-                    )}
+                    <SequenceInput
+                      value={this.props.value}
+                      ref={this._editorRef}
+                      onChecksChange={(tests) => {
+                        this.setState({ sequenceChecks: tests });
+                      }}
+                    />
                   </div>
                 </div>
 
@@ -708,7 +355,8 @@ export class IPScanSearch extends PureComponent /*:: <Props, State> */ {
                       <Button
                         className={f({
                           hidden:
-                            valid || !editorState.getCurrentContent().hasText(),
+                            this.state.sequenceChecks.valid ||
+                            !this.state.sequenceChecks.hasText,
                         })}
                         onClick={this._cleanUp}
                         borderColor="var(--colors-alert-main)"
@@ -739,8 +387,13 @@ export class IPScanSearch extends PureComponent /*:: <Props, State> */ {
                   >
                     <Button
                       submit
-                      className={f({ disabled: !valid })}
-                      disabled={tooShort || !valid}
+                      className={f({
+                        disabled: !this.state.sequenceChecks.valid,
+                      })}
+                      disabled={
+                        this.state.sequenceChecks.tooShort ||
+                        !this.state.sequenceChecks.valid
+                      }
                       value="Search"
                     >
                       Search
