+++ src/components/IPScan/Search/style.css
@@ -9,6 +9,10 @@
   font-size: 0.8rem;
   overflow: auto;
   margin-bottom: 0.5rem;
+  background-color: white;
+  border: 2px solid var(--colors-gray);
+  padding: 0.5rem;
+  cursor: text;
 
   & :global(.public-DraftEditorPlaceholder-root) {
     color: #cacaca;
@@ -90,6 +94,10 @@ div.line-with-buttons {
 
 .comment {
   color: gray;
+}
+.fasta-header {
+  color: rgb(11, 11, 54);
+  font-weight: bold;
 }
 
 .invalid-letter,
